+++ pl_bolts/models/detection/__init__.py
@@ -1,5 +1,7 @@
+from pl_bolts.models.detection import components  # noqa: F401
 from pl_bolts.models.detection.faster_rcnn import FasterRCNN  # noqa: F401
 
 __all__ = [
+    "components",
     "FasterRCNN",
 ]