+++ pl_bolts/models/gans/__init__.py
@@ -1,5 +1,7 @@
 from pl_bolts.models.gans.basic.basic_gan_module import GAN  # noqa: F401
+from pl_bolts.models.gans.dcgan.dcgan_module import DCGAN  # noqa: F401
 
 __all__ = [
     "GAN",
+    "DCGAN",
 ]